refactor the way to load the shemas to have something more generic
