+++ internal/api/interface/v1/dashboard/interface.go
@@ -15,7 +15,6 @@ package dashboard
 
 import (
 	"github.com/perses/common/etcd"
-	"github.com/perses/perses/internal/api/impl/v1/dashboard/schemas"
 	"github.com/perses/perses/internal/api/shared"
 	v1 "github.com/perses/perses/pkg/model/api/v1"
 )
@@ -44,5 +43,4 @@ type DAO interface {
 
 type Service interface {
 	shared.ToolboxService
-	GetValidator() schemas.Validator
 }
