+++ internal/api/shared/dependency/service_manager.go
@@ -31,6 +31,7 @@ import (
 	"github.com/perses/perses/internal/api/interface/v1/health"
 	"github.com/perses/perses/internal/api/interface/v1/project"
 	"github.com/perses/perses/internal/api/interface/v1/user"
+	"github.com/perses/perses/internal/api/shared/schemas"
 )
 
 type ServiceManager interface {
@@ -40,6 +41,7 @@ type ServiceManager interface {
 	GetGlobalDatasource() globaldatasource.Service
 	GetHealth() health.Service
 	GetProject() project.Service
+	GetSchemas() schemas.Schemas
 	GetUser() user.Service
 }
 
@@ -51,17 +53,19 @@ type service struct {
 	globalDatasource globaldatasource.Service
 	health           health.Service
 	project          project.Service
+	schemas          schemas.Schemas
 	user             user.Service
 }
 
 func NewServiceManager(dao PersistenceManager, conf config.Config) ServiceManager {
-	dashboardService := dashboardImpl.NewService(dao.GetDashboard(), conf)
 	datasourceService := datasourceImpl.NewService(dao.GetDatasource())
 	folderService := folderImpl.NewService(dao.GetFolder())
 	globalDatasourceService := globalDatasourceImpl.NewService(dao.GetGlobalDatasource())
 	healthService := healthImpl.NewService(dao.GetHealth())
 	projectService := projectImpl.NewService(dao.GetProject())
+	schemasService := schemas.New(conf.Schemas)
 	userService := userImpl.NewService(dao.GetUser())
+	dashboardService := dashboardImpl.NewService(dao.GetDashboard(), schemasService)
 	return &service{
 		dashboard:        dashboardService,
 		datasource:       datasourceService,
@@ -69,6 +73,7 @@ func NewServiceManager(dao PersistenceManager, conf config.Config) ServiceManage
 		globalDatasource: globalDatasourceService,
 		health:           healthService,
 		project:          projectService,
+		schemas:          schemasService,
 		user:             userService,
 	}
 }
@@ -95,6 +100,10 @@ func (s *service) GetHealth() health.Service {
 
 func (s *service) GetProject() project.Service {
 	return s.project
+}
+
+func (s *service) GetSchemas() schemas.Schemas {
+	return s.schemas
 }
 
 func (s *service) GetUser() user.Service {
