+++ internal/api/shared/schemas/base_def_panel.cue
