+++ internal/api/shared/schemas/base_def_query.cue
