+++ internal/api/shared/schemas/testdata/panels/first/panel.cue
