+++ internal/api/shared/schemas/testdata/panels/second/panel.cue
