+++ internal/api/shared/schemas/testdata/panels/third/panel.cue
