+++ internal/cli/cmd/lint/lint.go
@@ -18,7 +18,7 @@ import (
 	"io"
 
 	"github.com/perses/perses/internal/api/config"
-	"github.com/perses/perses/internal/api/impl/v1/dashboard/schemas"
+	"github.com/perses/perses/internal/api/shared/schemas"
 	"github.com/perses/perses/internal/cli/cmd"
 	"github.com/perses/perses/internal/cli/file"
 	"github.com/perses/perses/internal/cli/opt"
@@ -34,7 +34,7 @@ type option struct {
 	writer         io.Writer
 	chartsSchemas  string
 	queriesSchemas string
-	validator      schemas.Validator
+	sch            schemas.Schemas
 }
 
 func (o *option) Complete(args []string) error {
@@ -42,7 +42,7 @@ func (o *option) Complete(args []string) error {
 		return fmt.Errorf("no args are supported by the command 'lint'")
 	}
 	if len(o.chartsSchemas) > 0 && len(o.queriesSchemas) > 0 {
-		o.validator = schemas.NewValidator(config.Schemas{
+		o.sch = schemas.New(config.Schemas{
 			PanelsPath:  o.chartsSchemas,
 			QueriesPath: o.queriesSchemas,
 		})
@@ -70,13 +70,13 @@ func (o *option) SetWriter(writer io.Writer) {
 }
 
 func (o *option) validate(objects []modelAPI.Entity) error {
-	if o.validator == nil {
+	if o.sch == nil {
 		return nil
 	}
 	for _, object := range objects {
 		entity, ok := object.(*modelV1.Dashboard)
 		if ok {
-			if err := o.validator.Validate(entity.Spec.Panels); err != nil {
+			if err := o.sch.ValidatePanels(entity.Spec.Panels); err != nil {
 				return err
 			}
 		}
